✨ feature : 관리자 - 인플루언서 리뷰 기능 구현
- 관리자 - 인플루언서 리뷰 가져오기 기능 구현
- 관리자 - 인플루언서 리뷰 검색 기능 구현
- 관리자 - 인플루언서 리뷰 공개/비공개 기능 구현
- Service,Repository Test 완료
- Controller PostMan Test 완료
- 프론트엔드 스프링 변경으로 인하여 js 일부 수정

+++ admin-integration-service/src/main/java/com/linki/admin_integration_service/domain/operations/dto/InfluencerResponseDTO.java
@@ -2,6 +2,7 @@ package com.linki.admin_integration_service.domain.operations.dto;
 
 import lombok.Data;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 @Data
@@ -10,7 +11,7 @@ public class InfluencerResponseDTO {
     private String influencer;
     private String writer;
     private String review;
-    private double rating;
+    private BigDecimal rating;
     private LocalDateTime reviewDate;
     private boolean visibility;
     private String contractId;
