+++ admin-integration-service/src/main/java/com/linki/admin_integration_service/domain/operations/dto/InfluencerReviewDTO.java
@@ -1,7 +1,9 @@
 package com.linki.admin_integration_service.domain.operations.dto;
 
+
 import lombok.Data;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 @Data
@@ -10,7 +12,7 @@ public class InfluencerReviewDTO {
     private String influencer;
     private String writer;
     private String review;
-    private double rating;
+    private BigDecimal rating;
     private LocalDateTime reviewDate;
     private boolean visibility;
     private String contractId;
